Create fuel_new.py to enable conflict resolver
In target-base branch, the fuel_new.py file was deleted because it is identical to fuel.py. Now, when target-base is ready to be merged back into master (#46), there are conflicts of the target files, because the whole structure of the targets was changed. This can be usually solved with the online conflict resolver, but because fuel_new.py was deleted, it is impossible right now. So, I created this file temporarily and will delete it ASAP.